requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
asttokens==2.1.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cond==1.2.5
cryptography==38.0.4
cssselect==1.2.0
cv==1.0.0
debugpy==1.6.3
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
face-recognition-models==0.3.0
fake-useragent==1.1.1
frozenlist==1.3.3
future==0.18.2
GoogleScraper==0.2.4
greenlet==2.0.2
h11==0.14.0
idna==3.4
ifcfg==0.23
ipinfo==4.4.2
ipykernel==6.17.1
ipython==8.6.0
ipywidgets==8.0.2
jedi==0.18.1
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-widgets==3.0.3
keyboard==0.13.5
lxml==4.9.2
matplotlib-inline==0.1.6
MouseInfo==0.1.3
multidict==6.0.4
Naked==0.1.32
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.4
opencv-python==4.7.0.68
openpyxl==3.1.1
outcome==1.2.0
packaging==21.3
pandas==1.5.3
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
progressbar==2.5
prompt-toolkit==3.0.32
psutil==5.9.3
pure-eval==0.2.2
pvrecorder==1.1.1
py-cpuinfo==9.0.0
pyasn1==0.4.8
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.17
pycryptoplus==0.0.1
pygame==2.1.3.dev8
PyGetWindow==0.0.9
Pygments==2.13.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyMsgBox==1.0.9
PyMySQL==1.0.2
pynput==1.7.6
pyparsing==3.0.9
PyPDF2==2.11.2
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-whois==0.8.0
pytweening==1.0.4
pytz==2022.7.1
pyuser-agent==0.1.0
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rsa==4.9
scapy==2.5.0
scipy==1.10.1
selenium==4.8.0
shellescape==3.8.1
SimpleCV==1.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.4
sp==1.1.0
SpeechRecognition==3.9.0
SQLAlchemy==2.0.2
stack-data==0.6.1
style==1.1.0
timer==0.2.2
tkcalendar==1.6.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
update==0.0.1
urllib3==1.26.13
useragent==0.1.1
wavio==0.0.7
wcwidth==0.2.5
websocket-server==0.6.4
whois==0.9.27
widgetsnbextension==4.0.3
WMI==1.5.1
wsproto==1.2.0
yarl==1.8.2
